Implemented expandable text field
Created new custom edit text class that allows the original edit text to be covered with a larger text field when clicked. This creates the expanding effect. When the soft keyboard is hidden, the custom collapsable text field is also hidden to give the illusion of the text field collapsing. Challenges included how to detect the soft keyboard state, and since there was no built in function that allowed us to do that, we had to work around it by implementing the onKeyPreIme method in the custom edit text.